Create an index across all letters of all words in the game #91 gained a 18x speed increase for creating a rectangle.

+++ Sources/CrozzleRectangleMaker/Models/LetterIndexItemModel.swift
@@ -9,9 +9,9 @@ import Foundation
 /// Used to store the exact positions of a particular letter in each of the words of the game
 public struct LetterIndexItemModel {
     /// The position of the word in the sequence
-    let wordId: UInt8
+    let wordId: Int
     /// The position of the letter from the start of the word
-    let fromStartPos: UInt8
+    let fromStartPos: Int
     /// The position of the letter from the end of the word
     let fromEndPos: UInt8
 }
